feature(ramp): add deeplink folder and handler
