+++ app/core/DeeplinkManager/Handlers/handleRampUrl.ts
@@ -1,24 +1,1 @@
-import { NavigationProp, ParamListBase } from '@react-navigation/native';
-import { RampType } from '../../../reducers/fiatOrders/types';
-import Routes from '../../../constants/navigation/Routes';
-
-interface RampUrlOptions {
-  rampPath: string;
-  rampType: RampType;
-  navigation: NavigationProp<ParamListBase>;
-}
-
-export default function handleRampUrl({
-  rampPath: _rampPath,
-  rampType,
-  navigation,
-}: RampUrlOptions) {
-  switch (rampType) {
-    case RampType.BUY:
-      navigation.navigate(Routes.RAMP.BUY);
-      break;
-    case RampType.SELL:
-      navigation.navigate(Routes.RAMP.SELL);
-      break;
-  }
-}
+export { default } from '../../../components/UI/Ramp/deeplink/handleRampUrl';
